requirements.txt
Django==4.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary
PyJWT==2.8.0
faker